% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAS-package.R
\docType{package}
\name{FAS-package}
\alias{FAS-package}
\alias{FAS}
\title{FAS}
\description{
Bootstrap methods for selecting the tuning parameter for LASSO-type regression models and testing sparse regression coefficients
}
\author{
Jonas Striaukas (maintainer) \email{jonas.striaukas@gmail.com}
}
\keyword{package}
